oops, merging (#16)
* feat: initial token auth workflows

* rename from home to template

* fix token auth and rename bugs

* add extra guard test case

* move api controllers to using token auth

* update readme

* refactor: require password to delete mfa

Closes #5

* add middleware to fix proxy ip headers

* feat: add per user rate limits to example alert controller

* feat: add automatic handling of 429's and some 500 status codes

#10

+++ template/controllers/auth_controller.py
@@ -505,6 +505,20 @@ class AuthController(Controller):
 
     @post("/mfa/totp/delete", name="mfa_totp_delete", middleware=[EnsureAuth])
     async def totp_mfa_delete(self, request: Request) -> Template | Redirect:
+        form = await request.form()
+        password = form.get("password")
+        algorithm, iterations_, salt, hashed = Users.split_stored_password(
+            request.user.password
+        )
+        iterations = int(iterations_)
+        if not Users.hash_password(password, salt, iterations) == request.user.password:
+            alert(
+                request,
+                "Incorrect password, unable to delete MFA at this time",
+                level="error",
+            )
+            return Redirect(request.url_for("manage_totp_mfa"))
+
         # SOFT delete, but safe to stack for same user
         await constants.MFA_TOTP_PROVIDER.delete_registration(user=request.user)
         alert(
